feat(testing): add in check for if first arg is 'node'
If the first argument for `process.argv` is not `node` it means
that what is being passed in are the arguments for the CLI to test.
The `CommandTestFactory` will then prepend `node` and a random
file name for the `commander` package to see and deal with. This
will make sure it does not try to test again `jest` via the file
name argument.