fix: re-throw error regardless if it is a reading undefined error
Due to the need to call the custom error handler when it is passed
it makes more sense to generally just re-throw the existing error
if one occurs. So now, there will be a logged message about possibly
using a request scoped service and the error will be fully handled
by the error handler.

ref: #1039